refactor: rename scheduler to snake_case
